[Backport 5.1]: grpc: moving IsGRPCEnabled to resolve import cycle issue (#55362) (#55370)
Moving IsGRPCEnabled from internal/grpc to internal/conf package to
resolve the import cycle issue caused when using IsGRPCEnabled in the
/internal/api/internalapi package.

This issues occurs because internalapi depends on -> internal/grpc
(IsGRPCEnabled) depends on -> internal/conf -> which depends on
internalapi
