+++ internal/conf/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "computed.go",
         "conf.go",
         "diff.go",
+        "grpc.go",
         "helpers.go",
         "init.go",
         "log_sinks.go",
@@ -57,6 +58,7 @@ go_test(
         "client_test.go",
         "computed_test.go",
         "diff_test.go",
+        "grpc_test.go",
         "mocks_test.go",
         "validate_test.go",
     ],
