+++ internal/gitserver/BUILD.bazel
@@ -29,7 +29,6 @@ go_library(
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
         "//internal/gitserver/v1:gitserver",
-        "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/grpc/streamio",
         "//internal/honey",
