+++ internal/grpc/grpc.go
@@ -2,17 +2,12 @@
 package grpc
 
 import (
-	"context"
 	"net/http"
-	"os"
-	"strconv"
 	"strings"
 
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 	"google.golang.org/grpc"
-
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 )
 
 // MultiplexHandlers takes a gRPC server and a plain HTTP handler and multiplexes the
@@ -31,13 +26,4 @@ func MultiplexHandlers(grpcServer *grpc.Server, httpHandler http.Handler) http.H
 	// basically HTTP2 without TLS. The standard library does not implement the
 	// h2s protocol, so this hijacks h2s requests and handles them correctly.
 	return h2c.NewHandler(newHandler, &http2.Server{})
-}
-
-const envGRPCEnabled = "SG_FEATURE_FLAG_GRPC"
-
-func IsGRPCEnabled(ctx context.Context) bool {
-	if val, err := strconv.ParseBool(os.Getenv(envGRPCEnabled)); err == nil {
-		return val
-	}
-	return conf.Get().ExperimentalFeatures.EnableGRPC
 }
