+++ internal/search/searcher/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "//internal/endpoint",
         "//internal/errcode",
         "//internal/gitserver",
-        "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/httpcli",
         "//internal/limiter",
